Fixed a Bug / Technicality
Technically, when the board is expanding, it retains the *previous* left / right walls as Ghost Blocks, before adding new Ghost Blocks to each side.

Now, these *previous* left / right walls convert to Empty when moving to the new array, and the final process still overrides the *new* left / right walls to be Ghost.